+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Removed deprecated `AnalysisOptions.from` use `AnalysisOptions.fromMap` instead
 - Removed deprecated `Config.linesOfCodeWarningLevel` use `Config.linesOfExecutableCodeWarningLevel` instead
+- Add static code anti-patterns long-parameter-list
 
 # 1.10.0